move language options to separate file
