+++ src/components/LanguageButton/LanguageConstants.js
@@ -1,0 +1,25 @@
+import UsFlag from '../../assets/img/us-flag.svg';
+import IraqFlag from '../../assets/img/iraq-flag.svg';
+
+const languages = [
+  {
+    name: 'English',
+    code: 'en',
+    flag: UsFlag,
+    direction: 'ltr',
+  },
+  {
+    name: 'العربية',
+    code: 'ar',
+    flag: IraqFlag,
+    direction: 'rtl',
+  },
+  {
+    name: 'کوردی',
+    code: 'ku',
+    flag: 'https://upload.wikimedia.org/wikipedia/commons/3/35/Flag_of_Kurdistan.svg',
+    direction: 'rtl',
+  },
+];
+
+export default languages;
